+++ modules/azure-databricks-security-groups/main.tf
@@ -1,8 +1,8 @@
 // Create the security groups and make the associations with the subnets
 resource "azurerm_network_security_group" "databricks-subnet-private-sg" {
-  name                = "${var.security-group-name-prefix}-private-subnet-security-group"
+  name                = "${var.security_group_name_prefix}-private-subnet-security-group"
   location            = var.location
-  resource_group_name = var.virtual_network_rg_name
+  resource_group_name = var.vnet_resource_group_name
 
   tags = var.tags
 }
@@ -13,9 +13,9 @@ resource "azurerm_subnet_network_security_group_association" "private-subnet-sg-
 }
 
 resource "azurerm_network_security_group" "databricks-subnet-public-sg" {
-  name                = "${var.security-group-name-prefix}-public-subnet-security-group"
+  name                = "${var.security_group_name_prefix}-public-subnet-security-group"
   location            = var.location
-  resource_group_name = var.virtual_network_rg_name
+  resource_group_name = var.vnet_resource_group_name
 
   tags = var.tags
 }