+++ modules/azure-databricks-security-groups/variables.tf
@@ -1,4 +1,4 @@
-variable "security-group-name-prefix" {
+variable "security_group_name_prefix" {
   type        = string
   description = "Prefix for the names of the security groups created by this module"
   default     = "tamr-databricks"
@@ -9,7 +9,7 @@ variable "location" {
   description = "Location in which Databricks will be deployed"
 }
 
-variable "virtual_network_rg_name" {
+variable "vnet_resource_group_name" {
   type        = string
   description = "Name of resource group which contains the virtual network and subnets"
 }