+++ modules/azure-databricks-workspace/main.tf
@@ -1,6 +1,6 @@
 // Create the databricks workspace
 resource "azurerm_databricks_workspace" "module-databricks" {
-  name                = var.name
+  name                = var.workspace_name
   resource_group_name = var.resource_group_name
   location            = var.location
   sku                 = "standard"
@@ -9,7 +9,7 @@ resource "azurerm_databricks_workspace" "module-databricks" {
     no_public_ip        = false
     private_subnet_name = var.private_subnet_name
     public_subnet_name  = var.public_subnet_name
-    virtual_network_id  = var.virtual_network_id
+    virtual_network_id  = var.vnet_id
   }
 
   tags = var.tags