+++ variables.tf
@@ -1,4 +1,4 @@
-variable "name" {
+variable "workspace_name" {
   type        = string
   description = "Name of Databricks workspace"
 }
@@ -13,12 +13,12 @@ variable "location" {
   description = "Location in which Databricks will be deployed"
 }
 
-variable "virtual_network_id" {
+variable "vnet_id" {
   type        = string
   description = "ID of existing virtual network into which Databricks will be deployed"
 }
 
-variable "virtual_network_rg_name" {
+variable "vnet_resource_group_name" {
   type        = string
   description = "Name of resource group which contains the virtual network and subnets"
 }